# topcoder/Opportunity Tracker Backend Development/submission-de839d2f-70e8-4fe1-af6b-ade6c50cd5dc/opportunity-tracker/docs/swagger.yml
openapi: 3.0.0
servers: []
info:
  version: '1.0'
  title: Hess Opportunities Backend API
  description: |
    Backend API for the Hess Opportunities application.
...
paths:
  /opportunities:
    get:
      summary: Retrieve opportunities.  Filtering and pagination are supported.
      description: |
        Retrieve opportunities.  Filtering, pagination, and sorting are supported.
        <br><b>Authorization:</b>All logged-in users are allowed to perform this operation.
        <br><b>Implementation Notes:</b>
        <ul>
          <li>The main table to use is opportunity table.</li>
          <li>The opportunity_phase without an end_date will be the current status of an opportunity.</li>
          <li>productName is opportunity.name</li>
        </ul>
      tags:
        - Opportunities
      operationId: searchOpportunities
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/perPage'
        - $ref: '#/components/parameters/sortBy'
        - $ref: '#/components/parameters/sortOrder'
        - in: query
          name: status
          description: Filter by status.
          schema:
            type: array
            items:
              $ref: '#/components/schemas/OpportunityStatus'
        - in: query
          name: productName
          description: Search/Filter by product name.
          schema:
            type: string
        - in: query
          name: companyName
          description: Search/Filter by company name.
          schema:
            type: string
        - in: query
          name: technologyTheme
          description: Search/Filter by technology theme.
          schema:
            type: array
            items:
              type: integer
        - in: query
          name: lastUpdatedStart
          description: Lower bound for filtering by lastUpdated attribute.
          schema:
            type: string
            format: date-time
        - in: query
          name: lastUpdatedEnd
          description: Upper bound for filtering by lastUpdated attribute.
          schema:
            type: string
            format: date-time
        - in: query
          name: owner
          description: Search/Filter by owner name.
          schema:
            type: string
        - in: query
          name: tags
          description: Search/Filter by tags.
          schema:
            type: array
            items:
              type: integer
      responses:
        '200':
          $ref: '#/components/responses/OpportunitySearchResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      summary: Create an opportunity with the given data.
      description: |
        Create an opportunity with the given data.
        <br><b>Authorization:</b>All logged-in users are allowed to perform this operation.
        <br><b>Implementation Notes:</b>
        <ul>
          <li>The main table to use is opportunity table.</li>
          <li>The opportunity_phase without an end_date will be the current status of an opportunity.</li>
          <li>productName is opportunity.name</li>
          <li>The initial phase for the opportunity should also be created and set to "UnderEvaluation".
        </ul>
      tags:
        - Opportunities
      operationId: createOpportunity
      security:
        - bearer: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrEditOpportunityRequest'
      responses:
        '200':
          $ref:  '#/components/responses/OpportunityDetailsWithIdResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /opportunities/{opportunityId}:
    parameters:
      - $ref: '#/components/parameters/OpportunityId'
    get:
      summary: Retrieve the opportunity details as well as associated phases and discussion with given opportunity id.
      description: |
        This method will retrieve the opportunity details as well as associated phases and discussion with
        given opportunity id.
      tags:
        - Opportunities
      operationId: getOpportunityCompleteById
      security:
        - bearer: []
      responses:
        '200':
          $ref:  '#/components/responses/OpportunityCompleteResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      summary: Delete the given opportunity.
      description: |
        This method will delete the opportunity with given id. Only the owner is allowed to delete the opportunity.
      tags:
        - Opportunities
      operationId: deleteOpportunity
      security:
        - bearer: []
      responses:
        '204':
          $ref:  '#/components/responses/NoContent'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /opportunities/{opportunityId}/details:
    parameters:
      - $ref: '#/components/parameters/OpportunityId'
    get:
      summary: Retrieve the opportunity details for given opportunity id.
      description: |
        This method will retrieve the opportunity details for given opportunity id.
      tags:
        - Opportunities
      operationId: getOpportunityDetailsById
      security:
        - bearer: []
      responses:
        '200':
          $ref:  '#/components/responses/OpportunityDetailsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      summary: Edit the opportunity details for given opportunity.
      description: |
        This method will edit the opportunity details for the given opportunity.  Only the owner is allowed to perform
        this operation.
      tags:
        - Opportunities
      operationId: editOpportunityById
      security:
        - bearer: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrEditOpportunityRequest'
      responses:
        '200':
          $ref:  '#/components/responses/OpportunityDetailsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /opportunities/{opportunityId}/phases:
    parameters:
      - $ref: '#/components/parameters/OpportunityId'
    get:
      summary: Retrieve the opportunity phases for given opportunity id.
      description: |
        This method will retrieve the opportunity phases for given opportunity id.
      tags:
        - Opportunities
      operationId: getOpportunityPhasesById
      security:
        - bearer: []
      responses:
        '200':
          $ref: '#/components/responses/OpportunityPhaseArrayResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      summary: Change the active phase of the opportunity.
      description: |
        Change the active phase of the opportunity.  This will create a new phase as
        the new active phase of the opportunity and automatically end the previous phase.
        All existing phases will be returned by the method after the new phase has been
        added.
      tags:
        - Opportunities
      operationId: addOpportunityPhase
      security:
        - bearer: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddPhaseRequest'
      responses:
        '200':
          $ref: '#/components/responses/OpportunityPhaseArrayResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /opportunities/{opportunityId}/members:
    parameters:
      - $ref: '#/components/parameters/OpportunityId'
    get:
      summary: Retrieve the team members for given opportunity id.
      description: |
        This method will retrieve the team members for given opportunity id.
      tags:
        - Opportunities
      operationId: getOpportunityMembersById
      security:
        - bearer: []
      responses:
        '200':
          $ref: '#/components/responses/OpportunityTeamMembersResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      summary: Update the list of team members for given opportunity id.
      description: |
        This method will update the team members for given opportunity id.  Only the owner of the
        opportunity is allowed to perform this action.
      tags:
        - Opportunities
      operationId: updateOpportunityMembersById
      security:
        - bearer: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOpportunityMembersRequest'
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /opportunities/{opportunityId}/discussionPosts:
    parameters:
      - $ref: '#/components/parameters/OpportunityId'
    get:
      summary: Retrieve the discussion posts for given opportunity id.
      description: |
        This method will retrieve the discussion posts for given opportunity id.  Only the top-level
        posts are returned, and further drilling down will be required if there is a need to access
        a specific discussion thread.
      tags:
        - Opportunities
        - Discussions
      operationId: getOpportunityDiscussionsById
      security:
        - bearer: []
      responses:
        '200':
          $ref:  '#/components/responses/DiscussionPostArrayResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      summary: Add a new top-level discussion post for given opportunity id.
      description: |
        This method add a new top-level discussion post for given opportunity id.
      tags:
        - Opportunities
        - Discussions
      operationId: addTopLevelDiscussionPost
      security:
        - bearer: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostRequest'
      responses:
        '201':
          $ref:  '#/components/responses/NoContent'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /discussionPosts/{discussionPostId}/replies:
    parameters:
      - $ref: '#/components/parameters/DiscussionPostId'
    get:
      summary: Retrieve the replies for the given discussion post.
      description: Retrieve the replies for the given discussion post.
      tags:
        - Discussions
      operationId: getDiscussionPostReplies
      security:
        - bearer: []
      responses:
        '200':
          $ref: '#/components/responses/DiscussionPostArrayResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      summary: Post a reply for the given discussion post.
      description: Post a reply for the given discussion post.
      tags:
        - Discussions
      operationId: createDiscussionReply
      security:
        - bearer: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostRequest'
      responses:
        '201':
          $ref: '#/components/responses/NoContent'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    page:
      name: page
      in: query
      description: The page number. 1 based.
      required: false
      schema:
        type: integer
        default: 1
    perPage:
      name: perPage
      in: query
      description: The number of entities shown in one page.
      required: false
      schema:
        type: integer
        default: 10
        maximum: 100
    sortBy:
      name: sortBy
      in: query
      description: The field used to sort the results
      required: false
      schema:
        type: string
        enum: ['ProductName', 'Company', 'TechTheme', 'Status', 'LastUpdated', 'Owner', 'Tags', 'Views']
    sortOrder:
      name: sortOrder
      in: query
      description: The sorting order (asc/desc).
      required: false
      schema:
        type: string
        enum:
          - asc
          - desc
        default: desc
    OpportunityId:
      name: opportunityId
      in: path
      required: true
      description: The opportunity id
      schema:
        type: integer
    DiscussionPostId:
      name: discussionPostId
      in: path
      required: true
      description: The discussion post id
      schema:
        type: integer
...
  schemas:
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
...
    AddPhaseRequest:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/OpportunityStatus'
...
    UpdateOpportunityMembersRequest:
      type: object
      properties:
        members:
          type: array
          description: A list of user ids that will be members of the given opportunity.
          items:
            type: integer
    CreatePostRequest:
      type: object
      properties:
        text:
          type: string
          description: The text content of the post.
    CreateOrEditOpportunityRequest:
      type: object
      properties:
        name:
          type: string
          description: The name of the opportunity.
        technologyTheme:
          type: integer
          description: The technology theme id.
        company:
          type: string
          description: The name of the company associated with this opportunity.
        source:
          type: integer
          description: The source id of the opportunity.
        description:
          type: string
          description: The opportunity description.
        tags:
          type: array
          description: An array of tags associated with the opportunity.
          items:
            type: integer
        supportingDocuments:
          type: array
          description: Array pointing to ids of uploaded documents.
          items:
            type: number
        usefulLinks:
          type: array
          description: An array of useful links associated with the opportunity.
          items:
            type: string
    OpportunityPhase:
      type: object
      properties:
        id:
          type: number
          description: The id of the phase.
        status:
          type: string
          description: The status of the phase.
        startDate:
          type: string
          format: date-time
          description: The start date of the phase.  If the phase has not yet started, this is null.
        endDate:
          type: string
          format: date-time
          description: The end date of the phase.  If the phase has not yet started, or has not completed, this is null.
        isActive:
          type: boolean
          description: |
            Indicates whether this is the currently active phase of the opportunity it is associated with.
...
    UserShort:
      type: object
      properties:
        id:
          type: number
          description: The id of the user.
        name:
          type: string
          description: The name of the user.
        profilePic:
          type: string
          description: The profile pic of the user.
    OpportunityComplete:
      type: object
      properties:
        id:
          type: integer
          description: The id of the opportunity.
        ownerId:
          type: integer
          description: The id of the opportunity owner.
        opportunityDetails:
          $ref: '#/components/schemas/OpportunityDetails'
        opportunityPhases:
          type: array
          items:
            $ref: '#/components/schemas/OpportunityPhase'
        discussions:
          type: array
          items:
            $ref: '#/components/schemas/DiscussionPost'
    OpportunityDetailsWithId:
      allOf:
        - type: object
          properties:
            id:
              type: integer
              description: The id of the opportunity.
        - $ref: '#/components/schemas/OpportunityDetails'
    OpportunityDetails:
      type: object
      properties:
        name:
          type: string
          description: The name of the opportunity.
        updatedOn:
          type: string
          format: date-time
          description: When the opportunity was last updated.
        technologyTheme:
          $ref: '#/components/schemas/TechnologyTheme'
        company:
          type: string
          description: The name of the company.
        source:
          $ref: '#/components/schemas/OpportunitySource'
        description:
          type: string
          description: The description of the opportunity.
        tags:
          type: array
          description: The tags associated with the opportunity.
          items:
            type: string
        documents:
          type: array
          description: An array of documents associated with the opportunity.
          items:
            $ref: '#/components/schemas/OpportunityDocument'
        links:
          type: array
          items:
            type: string
    OpportunityDocument:
      type: object
      properties:
        name:
          type: string
          description: The name of the document.
        documentType:
          type: string
          description: The MIME document type.
        location:
          type: string
          description: A link to access the document.
    DiscussionPost:
      type: object
      properties:
        author:
          $ref: '#/components/schemas/UserShort'
        date:
          type: string
          format: date-time
          description: The date of the post.
        text:
          type: string
          description: The text of the post.
        replyCount:
          type: number
          description: The number of replies to this post.
    TechnologyTheme:
      type: object
      properties:
        icon:
          type: string
          description: The icon to use for the theme.
        name:
          type: string
          description: The name of the theme.
    Tag:
      type: object
      properties:
        id:
          type: integer
          description: The id of the tag.
        label:
          type: string
          description: The label of the tag.
    OpportunitySource:
      type: object
      properties:
        id:
          type: integer
          description: The id of the source.
        name:
          type: string
          description: The name of the source.
    OpportunityStatus:
      type: string
      enum: [ 'UnderEvaluation', 'EvaluationPass', 'EvaluationFail', 'InPilotTest', 'PilotTestPass',
                'PilotTestFail', 'Implemented' ]
...
    OpportunityTeamMembers:
      type: object
      description: The team members of a given opportunity.
      properties:
        owner:
          $ref: '#/components/schemas/UserShort'
        members:
          type: array
          items:
            $ref: '#/components/schemas/UserShort'
...
    OpportunitySearchResultItem:
      type: object
      properties:
        id:
          type: integer
          description: The id of the opportunity.
        description:
          type: string
          description: The description of the opportunity.
        name:
          type: string
          description: The name of the opportunity.
        company:
          type: string
          description: The company of the opportunity.
        technologyTheme:
          $ref: '#/components/schemas/TechnologyTheme'
        status:
          $ref: '#/components/schemas/OpportunityStatus'
        lastUpdatedOn:
          type: string
          description: When the opportunity was last updated.
          format: date-time
        owner:
          $ref: '#/components/schemas/UserShort'
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        views:
          type: integer
          description: The number of views for this opportunity.
  headers:
    X-Page:
      schema:
        type: integer
      description: Index of the current page (starting at 1)
    X-Per-Page:
      schema:
        type: integer
      description: Number of items per page
    X-Prev-Page:
      schema:
        type: integer
      description: Index of the previous page
    X-Next-Page:
      schema:
        type: integer
      description: Index of the next page
    X-Total:
      schema:
        type: integer
      description: Total number of items
    X-Total-Pages:
      schema:
        type: integer
      description: Total number of pages
...
  responses:
    OpportunitySearchResponse:
      description: Response that for opportunity search.
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/OpportunitySearchResultItem'
      headers:
        X-Page:
          $ref: '#/components/headers/X-Page'
        X-Per-Page:
          $ref: '#/components/headers/X-Per-Page'
        X-Next-Page:
          $ref: '#/components/headers/X-Next-Page'
        X-Prev-Page:
          $ref: '#/components/headers/X-Prev-Page'
        X-Total:
          $ref: '#/components/headers/X-Total'
        X-Total-Pages:
          $ref: '#/components/headers/X-Total-Pages'
...
    OpportunityTeamMembersResponse:
      description: Response that provides an OpportunityTeamMembers object.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/OpportunityTeamMembers'
...
    OpportunityCompleteResponse:
      description: Response that provides an OpportunityComplete object.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/OpportunityComplete'
...
    OpportunityDetailsResponse:
      description: Response that provides an OpportunityDetails object.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/OpportunityDetails'
    OpportunityDetailsWithIdResponse:
      description: Response that provides an OpportunityDetailsWithId object.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/OpportunityDetailsWithId'
    OpportunityPhaseArrayResponse:
      description: Response that provides an array of OpportunityPhase objects.
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/OpportunityPhase'
    DiscussionPostArrayResponse:
      description: Response that provides an array of DicussionPost objects.
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/DiscussionPost'
...
    NoContent:
      description: |
        The operation completed successfully but no content is needed
        in the response.
    BadRequest:
      description: Bad request - Problem with the request (E.g. Missing parameters)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized - Token is missing or Invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Not found - Requested entity is not found in database
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalServerError:
      description: Internal Server Error - Request is valid but operation failed at server side
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'